+++ lib/fl_downloader.dart
@@ -1,1 +1,2 @@
 export 'src/fl_downloader_src.dart';
+export 'src/storage_permission_status.dart';